requirements.txt
mdanalysis==2.7.0
pandas==2.2.1
numpy==1.26.0
matplotlib==3.8.2
scikit-learn==1.6.1
scipy==1.12.0
pdb-tools==2.5.0
biopython==1.85.0